+++ joblib/disk.py
@@ -46,10 +46,10 @@ def memstr_to_bytes(text):
     units = dict(K=kilo, M=kilo ** 2, G=kilo ** 3)
     try:
         size = int(units[text[-1]] * float(text[:-1]))
-    except (KeyError, ValueError):
+    except (KeyError, ValueError) as e:
         raise ValueError(
             "Invalid literal for size give: %s (type %s) should be "
-            "alike '10G', '500M', '50K'." % (text, type(text)))
+            "alike '10G', '500M', '50K'." % (text, type(text))) from e
     return size