+++ joblib/externals/cloudpickle/__init__.py
@@ -1,11 +1,11 @@
 from __future__ import absolute_import
 
-import sys
-import pickle
 
+from .cloudpickle import *  # noqa
+from .cloudpickle_fast import CloudPickler, dumps, dump  # noqa
 
-from .cloudpickle import *
-if sys.version_info[:2] >= (3, 8):
-    from .cloudpickle_fast import CloudPickler, dumps, dump
+# Conform to the convention used by python serialization libraries, which
+# expose their Pickler subclass at top-level under the  "Pickler" name.
+Pickler = CloudPickler
 
-__version__ = '1.4.1'
+__version__ = '1.6.0'